chore: implement ignore guard
